requirements.txt
altair~=5.2.0
jupyter==1.0.0
python-dotenv~=1.0.1
scikit-learn==1.4.0
seaborn==0.13.2
pygithub==2.2.0
streamlit~=1.31.0
pandas~=2.2.0
